Make content of bidirectional scroll more efficient and smoother when scrolling
Added slivers support to make large content smoother on scroll. I have
tried to add calculating the height and width if not provided however
it would be more efficient if they’re both passed as params if known.
Otherwise, it will try to do its best to figure the width and height